requirements.txt
fastapi>=0.95.0
fastapi[standard]
uvicorn[standard]>=0.22.0
pydantic>=1.10.0
pymongo>=4.3.0
PyJWT>=2.6.0
bcrypt>=4.0.0
python-dotenv>=0.21.0
pydantic-settings
azure-communication-email
openai
numpy
azure-storage-blob
requests
APScheduler
motor
azure-identity
pytz
httpx
httpx[http2]
beautifulsoup4
lxml
cachetools
python-dateutil

